==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_raw_data(year: int, month: int) -> Path:
    ''''''
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw/rides_{year}--{month:02d}.parquet'
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f'{URL} is not available')

download_raw_data(year =2024, month = 1)

# load file into a pandas dataframe

import pandas as pd
rides = pd.read_parquet('../data/raw/rides_2024--01.parquet')
rides.head()

rides = rides[['tpep_pickup_datetime','PULocationID']]

rides.info()

rides.rename(columns={'tpep_pickup_datetime': 'pickup_datetime',
                     'PULocationID': 'pickup_location_id',
                     }, inplace=True)

rides.head(20)

rides['pickup_datetime'].describe()

#remove dates not in January 2024
rides = rides[rides.pickup_datetime >= '2024-01-01']
rides = rides[rides.pickup_datetime < '2024-02-01']
rides['pickup_datetime'].describe()

# Store the validated file

rides.to_parquet('../data/transformed/validated_rides_2024_01.parquet')
